api/requirements.txt
psycopg2-binary>=2.9
asyncpg>=0.27
